refactor: 574 - move ListData type to separate file
