+++ src/widgets/study-path/ui/stage-card/stage-card.types.ts
@@ -1,4 +1,4 @@
-import { type ListData } from '@/shared/ui/list';
+import type { ListData, ListType } from '@/shared/types';
 
 export interface Stage {
   id: number | string;
@@ -16,5 +16,5 @@ export interface Stage {
 }
 
 export interface StageCardProps extends Stage {
-  marked?: boolean;
+  type?: ListType;
 }
